+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/ze_generated_example_msiximages_client_test.go
@@ -24,21 +24,20 @@ func ExampleMsixImagesClient_NewExpandPager() {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client, err := armdesktopvirtualization.NewMsixImagesClient("<subscription-id>", cred, nil)
+	client, err := armdesktopvirtualization.NewMsixImagesClient("daefabc0-95b4-48b3-b645-8a753a63c4fa", cred, nil)
 	if err != nil {
 		log.Fatalf("failed to create client: %v", err)
 	}
-	pager := client.NewExpandPager("<resource-group-name>",
-		"<host-pool-name>",
+	pager := client.NewExpandPager("resourceGroup1",
+		"hostpool1",
 		armdesktopvirtualization.MSIXImageURI{
-			URI: to.Ptr("<uri>"),
+			URI: to.Ptr("imagepath"),
 		},
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)
 		if err != nil {
 			log.Fatalf("failed to advance page: %v", err)
-			return
 		}
 		for _, v := range nextResult.Value {
 			// TODO: use page item