+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/ze_generated_example_startmenuitems_client_test.go
@@ -23,18 +23,17 @@ func ExampleStartMenuItemsClient_NewListPager() {
 		log.Fatalf("failed to obtain a credential: %v", err)
 	}
 	ctx := context.Background()
-	client, err := armdesktopvirtualization.NewStartMenuItemsClient("<subscription-id>", cred, nil)
+	client, err := armdesktopvirtualization.NewStartMenuItemsClient("daefabc0-95b4-48b3-b645-8a753a63c4fa", cred, nil)
 	if err != nil {
 		log.Fatalf("failed to create client: %v", err)
 	}
-	pager := client.NewListPager("<resource-group-name>",
-		"<application-group-name>",
+	pager := client.NewListPager("resourceGroup1",
+		"applicationGroup1",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)
 		if err != nil {
 			log.Fatalf("failed to advance page: %v", err)
-			return
 		}
 		for _, v := range nextResult.Value {
 			// TODO: use page item