+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/zz_generated_hostpools_client.go
@@ -39,7 +39,7 @@ func NewHostPoolsClient(subscriptionID string, credential azcore.TokenCredential
 	if options == nil {
 		options = &arm.ClientOptions{}
 	}
-	ep := cloud.AzurePublicCloud.Services[cloud.ResourceManager].Endpoint
+	ep := cloud.AzurePublic.Services[cloud.ResourceManager].Endpoint
 	if c, ok := options.Cloud.Services[cloud.ResourceManager]; ok {
 		ep = c.Endpoint
 	}
@@ -57,6 +57,7 @@ func NewHostPoolsClient(subscriptionID string, credential azcore.TokenCredential
 
 // CreateOrUpdate - Create or update a host pool.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // hostPoolName - The name of the host pool within the specified resource group
 // hostPool - Object containing HostPool definitions.
@@ -99,7 +100,7 @@ func (client *HostPoolsClient) createOrUpdateCreateRequest(ctx context.Context, 
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, hostPool)
 }
 
@@ -114,6 +115,7 @@ func (client *HostPoolsClient) createOrUpdateHandleResponse(resp *http.Response)
 
 // Delete - Remove a host pool.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // hostPoolName - The name of the host pool within the specified resource group
 // options - HostPoolsClientDeleteOptions contains the optional parameters for the HostPoolsClient.Delete method.
@@ -157,12 +159,13 @@ func (client *HostPoolsClient) deleteCreateRequest(ctx context.Context, resource
 		reqQP.Set("force", strconv.FormatBool(*options.Force))
 	}
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
 // Get - Get a host pool.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // hostPoolName - The name of the host pool within the specified resource group
 // options - HostPoolsClientGetOptions contains the optional parameters for the HostPoolsClient.Get method.
@@ -203,7 +206,7 @@ func (client *HostPoolsClient) getCreateRequest(ctx context.Context, resourceGro
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -218,9 +221,10 @@ func (client *HostPoolsClient) getHandleResponse(resp *http.Response) (HostPools
 
 // NewListPager - List hostPools in subscription.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // options - HostPoolsClientListOptions contains the optional parameters for the HostPoolsClient.List method.
 func (client *HostPoolsClient) NewListPager(options *HostPoolsClientListOptions) *runtime.Pager[HostPoolsClientListResponse] {
-	return runtime.NewPager(runtime.PageProcessor[HostPoolsClientListResponse]{
+	return runtime.NewPager(runtime.PagingHandler[HostPoolsClientListResponse]{
 		More: func(page HostPoolsClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -261,7 +265,7 @@ func (client *HostPoolsClient) listCreateRequest(ctx context.Context, options *H
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -276,11 +280,12 @@ func (client *HostPoolsClient) listHandleResponse(resp *http.Response) (HostPool
 
 // NewListByResourceGroupPager - List hostPools.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // options - HostPoolsClientListByResourceGroupOptions contains the optional parameters for the HostPoolsClient.ListByResourceGroup
 // method.
 func (client *HostPoolsClient) NewListByResourceGroupPager(resourceGroupName string, options *HostPoolsClientListByResourceGroupOptions) *runtime.Pager[HostPoolsClientListByResourceGroupResponse] {
-	return runtime.NewPager(runtime.PageProcessor[HostPoolsClientListByResourceGroupResponse]{
+	return runtime.NewPager(runtime.PagingHandler[HostPoolsClientListByResourceGroupResponse]{
 		More: func(page HostPoolsClientListByResourceGroupResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0
 		},
@@ -325,7 +330,7 @@ func (client *HostPoolsClient) listByResourceGroupCreateRequest(ctx context.Cont
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -340,6 +345,7 @@ func (client *HostPoolsClient) listByResourceGroupHandleResponse(resp *http.Resp
 
 // RetrieveRegistrationToken - Registration token of the host pool.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // hostPoolName - The name of the host pool within the specified resource group
 // options - HostPoolsClientRetrieveRegistrationTokenOptions contains the optional parameters for the HostPoolsClient.RetrieveRegistrationToken
@@ -381,7 +387,7 @@ func (client *HostPoolsClient) retrieveRegistrationTokenCreateRequest(ctx contex
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
 }
 
@@ -396,6 +402,7 @@ func (client *HostPoolsClient) retrieveRegistrationTokenHandleResponse(resp *htt
 
 // Update - Update a host pool.
 // If the operation fails it returns an *azcore.ResponseError type.
+// Generated from API version 2022-02-10-preview
 // resourceGroupName - The name of the resource group. The name is case insensitive.
 // hostPoolName - The name of the host pool within the specified resource group
 // options - HostPoolsClientUpdateOptions contains the optional parameters for the HostPoolsClient.Update method.
@@ -436,7 +443,7 @@ func (client *HostPoolsClient) updateCreateRequest(ctx context.Context, resource
 	reqQP := req.Raw().URL.Query()
 	reqQP.Set("api-version", "2022-02-10-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
-	req.Raw().Header.Set("Accept", "application/json")
+	req.Raw().Header["Accept"] = []string{"application/json"}
 	if options != nil && options.HostPool != nil {
 		return req, runtime.MarshalAsJSON(req, *options.HostPool)
 	}