+++ sdk/resourcemanager/desktopvirtualization/armdesktopvirtualization/zz_generated_models_serde.go
@@ -10,6 +10,7 @@ package armdesktopvirtualization
 
 import (
 	"encoding/json"
+	"fmt"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"reflect"
@@ -54,14 +55,6 @@ func (a ApplicationGroup) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ApplicationGroupList.
-func (a ApplicationGroupList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", a.NextLink)
-	populate(objectMap, "value", a.Value)
-	return json.Marshal(objectMap)
-}
-
 // MarshalJSON implements the json.Marshaller interface for type ApplicationGroupPatch.
 func (a ApplicationGroupPatch) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -70,14 +63,6 @@ func (a ApplicationGroupPatch) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "properties", a.Properties)
 	populate(objectMap, "tags", a.Tags)
 	populate(objectMap, "type", a.Type)
-	return json.Marshal(objectMap)
-}
-
-// MarshalJSON implements the json.Marshaller interface for type ApplicationList.
-func (a ApplicationList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", a.NextLink)
-	populate(objectMap, "value", a.Value)
 	return json.Marshal(objectMap)
 }
 
@@ -113,67 +98,59 @@ func (a ApplicationProperties) MarshalJSON() ([]byte, error) {
 func (a *ApplicationProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "applicationType":
-			err = unpopulate(val, &a.ApplicationType)
+			err = unpopulate(val, "ApplicationType", &a.ApplicationType)
 			delete(rawMsg, key)
 		case "commandLineArguments":
-			err = unpopulate(val, &a.CommandLineArguments)
+			err = unpopulate(val, "CommandLineArguments", &a.CommandLineArguments)
 			delete(rawMsg, key)
 		case "commandLineSetting":
-			err = unpopulate(val, &a.CommandLineSetting)
+			err = unpopulate(val, "CommandLineSetting", &a.CommandLineSetting)
 			delete(rawMsg, key)
 		case "description":
-			err = unpopulate(val, &a.Description)
+			err = unpopulate(val, "Description", &a.Description)
 			delete(rawMsg, key)
 		case "filePath":
-			err = unpopulate(val, &a.FilePath)
+			err = unpopulate(val, "FilePath", &a.FilePath)
 			delete(rawMsg, key)
 		case "friendlyName":
-			err = unpopulate(val, &a.FriendlyName)
+			err = unpopulate(val, "FriendlyName", &a.FriendlyName)
 			delete(rawMsg, key)
 		case "iconContent":
 			err = runtime.DecodeByteArray(string(val), &a.IconContent, runtime.Base64StdFormat)
 			delete(rawMsg, key)
 		case "iconHash":
-			err = unpopulate(val, &a.IconHash)
+			err = unpopulate(val, "IconHash", &a.IconHash)
 			delete(rawMsg, key)
 		case "iconIndex":
-			err = unpopulate(val, &a.IconIndex)
+			err = unpopulate(val, "IconIndex", &a.IconIndex)
 			delete(rawMsg, key)
 		case "iconPath":
-			err = unpopulate(val, &a.IconPath)
+			err = unpopulate(val, "IconPath", &a.IconPath)
 			delete(rawMsg, key)
 		case "msixPackageApplicationId":
-			err = unpopulate(val, &a.MsixPackageApplicationID)
+			err = unpopulate(val, "MsixPackageApplicationID", &a.MsixPackageApplicationID)
 			delete(rawMsg, key)
 		case "msixPackageFamilyName":
-			err = unpopulate(val, &a.MsixPackageFamilyName)
+			err = unpopulate(val, "MsixPackageFamilyName", &a.MsixPackageFamilyName)
 			delete(rawMsg, key)
 		case "objectId":
-			err = unpopulate(val, &a.ObjectID)
+			err = unpopulate(val, "ObjectID", &a.ObjectID)
 			delete(rawMsg, key)
 		case "showInPortal":
-			err = unpopulate(val, &a.ShowInPortal)
+			err = unpopulate(val, "ShowInPortal", &a.ShowInPortal)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type DesktopList.
-func (d DesktopList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", d.NextLink)
-	populate(objectMap, "value", d.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type DesktopPatch.
@@ -199,40 +176,32 @@ func (d DesktopProperties) MarshalJSON() ([]byte, error) {
 func (d *DesktopProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", d, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "description":
-			err = unpopulate(val, &d.Description)
+			err = unpopulate(val, "Description", &d.Description)
 			delete(rawMsg, key)
 		case "friendlyName":
-			err = unpopulate(val, &d.FriendlyName)
+			err = unpopulate(val, "FriendlyName", &d.FriendlyName)
 			delete(rawMsg, key)
 		case "iconContent":
 			err = runtime.DecodeByteArray(string(val), &d.IconContent, runtime.Base64StdFormat)
 			delete(rawMsg, key)
 		case "iconHash":
-			err = unpopulate(val, &d.IconHash)
+			err = unpopulate(val, "IconHash", &d.IconHash)
 			delete(rawMsg, key)
 		case "objectId":
-			err = unpopulate(val, &d.ObjectID)
+			err = unpopulate(val, "ObjectID", &d.ObjectID)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", d, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type ExpandMsixImageList.
-func (e ExpandMsixImageList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", e.NextLink)
-	populate(objectMap, "value", e.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type ExpandMsixImageProperties.
@@ -258,53 +227,53 @@ func (e ExpandMsixImageProperties) MarshalJSON() ([]byte, error) {
 func (e *ExpandMsixImageProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", e, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "displayName":
-			err = unpopulate(val, &e.DisplayName)
+			err = unpopulate(val, "DisplayName", &e.DisplayName)
 			delete(rawMsg, key)
 		case "imagePath":
-			err = unpopulate(val, &e.ImagePath)
+			err = unpopulate(val, "ImagePath", &e.ImagePath)
 			delete(rawMsg, key)
 		case "isActive":
-			err = unpopulate(val, &e.IsActive)
+			err = unpopulate(val, "IsActive", &e.IsActive)
 			delete(rawMsg, key)
 		case "isRegularRegistration":
-			err = unpopulate(val, &e.IsRegularRegistration)
+			err = unpopulate(val, "IsRegularRegistration", &e.IsRegularRegistration)
 			delete(rawMsg, key)
 		case "lastUpdated":
-			err = unpopulateTimeRFC3339(val, &e.LastUpdated)
+			err = unpopulateTimeRFC3339(val, "LastUpdated", &e.LastUpdated)
 			delete(rawMsg, key)
 		case "packageAlias":
-			err = unpopulate(val, &e.PackageAlias)
+			err = unpopulate(val, "PackageAlias", &e.PackageAlias)
 			delete(rawMsg, key)
 		case "packageApplications":
-			err = unpopulate(val, &e.PackageApplications)
+			err = unpopulate(val, "PackageApplications", &e.PackageApplications)
 			delete(rawMsg, key)
 		case "packageDependencies":
-			err = unpopulate(val, &e.PackageDependencies)
+			err = unpopulate(val, "PackageDependencies", &e.PackageDependencies)
 			delete(rawMsg, key)
 		case "packageFamilyName":
-			err = unpopulate(val, &e.PackageFamilyName)
+			err = unpopulate(val, "PackageFamilyName", &e.PackageFamilyName)
 			delete(rawMsg, key)
 		case "packageFullName":
-			err = unpopulate(val, &e.PackageFullName)
+			err = unpopulate(val, "PackageFullName", &e.PackageFullName)
 			delete(rawMsg, key)
 		case "packageName":
-			err = unpopulate(val, &e.PackageName)
+			err = unpopulate(val, "PackageName", &e.PackageName)
 			delete(rawMsg, key)
 		case "packageRelativePath":
-			err = unpopulate(val, &e.PackageRelativePath)
+			err = unpopulate(val, "PackageRelativePath", &e.PackageRelativePath)
 			delete(rawMsg, key)
 		case "version":
-			err = unpopulate(val, &e.Version)
+			err = unpopulate(val, "Version", &e.Version)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", e, err)
 		}
 	}
 	return nil
@@ -326,14 +295,6 @@ func (h HostPool) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "systemData", h.SystemData)
 	populate(objectMap, "tags", h.Tags)
 	populate(objectMap, "type", h.Type)
-	return json.Marshal(objectMap)
-}
-
-// MarshalJSON implements the json.Marshaller interface for type HostPoolList.
-func (h HostPoolList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", h.NextLink)
-	populate(objectMap, "value", h.Value)
 	return json.Marshal(objectMap)
 }
 
@@ -378,14 +339,6 @@ func (h HostPoolProperties) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// MarshalJSON implements the json.Marshaller interface for type MSIXPackageList.
-func (m MSIXPackageList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", m.NextLink)
-	populate(objectMap, "value", m.Value)
-	return json.Marshal(objectMap)
-}
-
 // MarshalJSON implements the json.Marshaller interface for type MSIXPackagePatch.
 func (m MSIXPackagePatch) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -417,47 +370,47 @@ func (m MSIXPackageProperties) MarshalJSON() ([]byte, error) {
 func (m *MSIXPackageProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "displayName":
-			err = unpopulate(val, &m.DisplayName)
+			err = unpopulate(val, "DisplayName", &m.DisplayName)
 			delete(rawMsg, key)
 		case "imagePath":
-			err = unpopulate(val, &m.ImagePath)
+			err = unpopulate(val, "ImagePath", &m.ImagePath)
 			delete(rawMsg, key)
 		case "isActive":
-			err = unpopulate(val, &m.IsActive)
+			err = unpopulate(val, "IsActive", &m.IsActive)
 			delete(rawMsg, key)
 		case "isRegularRegistration":
-			err = unpopulate(val, &m.IsRegularRegistration)
+			err = unpopulate(val, "IsRegularRegistration", &m.IsRegularRegistration)
 			delete(rawMsg, key)
 		case "lastUpdated":
-			err = unpopulateTimeRFC3339(val, &m.LastUpdated)
+			err = unpopulateTimeRFC3339(val, "LastUpdated", &m.LastUpdated)
 			delete(rawMsg, key)
 		case "packageApplications":
-			err = unpopulate(val, &m.PackageApplications)
+			err = unpopulate(val, "PackageApplications", &m.PackageApplications)
 			delete(rawMsg, key)
 		case "packageDependencies":
-			err = unpopulate(val, &m.PackageDependencies)
+			err = unpopulate(val, "PackageDependencies", &m.PackageDependencies)
 			delete(rawMsg, key)
 		case "packageFamilyName":
-			err = unpopulate(val, &m.PackageFamilyName)
+			err = unpopulate(val, "PackageFamilyName", &m.PackageFamilyName)
 			delete(rawMsg, key)
 		case "packageName":
-			err = unpopulate(val, &m.PackageName)
+			err = unpopulate(val, "PackageName", &m.PackageName)
 			delete(rawMsg, key)
 		case "packageRelativePath":
-			err = unpopulate(val, &m.PackageRelativePath)
+			err = unpopulate(val, "PackageRelativePath", &m.PackageRelativePath)
 			delete(rawMsg, key)
 		case "version":
-			err = unpopulate(val, &m.Version)
+			err = unpopulate(val, "Version", &m.Version)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
 		}
 	}
 	return nil
@@ -480,25 +433,25 @@ func (m MsixPackageApplications) MarshalJSON() ([]byte, error) {
 func (m *MsixPackageApplications) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", m, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "appId":
-			err = unpopulate(val, &m.AppID)
+			err = unpopulate(val, "AppID", &m.AppID)
 			delete(rawMsg, key)
 		case "appUserModelID":
-			err = unpopulate(val, &m.AppUserModelID)
+			err = unpopulate(val, "AppUserModelID", &m.AppUserModelID)
 			delete(rawMsg, key)
 		case "description":
-			err = unpopulate(val, &m.Description)
+			err = unpopulate(val, "Description", &m.Description)
 			delete(rawMsg, key)
 		case "friendlyName":
-			err = unpopulate(val, &m.FriendlyName)
+			err = unpopulate(val, "FriendlyName", &m.FriendlyName)
 			delete(rawMsg, key)
 		case "iconImageName":
-			err = unpopulate(val, &m.IconImageName)
+			err = unpopulate(val, "IconImageName", &m.IconImageName)
 			delete(rawMsg, key)
 		case "rawIcon":
 			err = runtime.DecodeByteArray(string(val), &m.RawIcon, runtime.Base64StdFormat)
@@ -508,26 +461,10 @@ func (m *MsixPackageApplications) UnmarshalJSON(data []byte) error {
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", m, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type PrivateEndpointConnectionListResultWithSystemData.
-func (p PrivateEndpointConnectionListResultWithSystemData) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", p.NextLink)
-	populate(objectMap, "value", p.Value)
-	return json.Marshal(objectMap)
-}
-
-// MarshalJSON implements the json.Marshaller interface for type PrivateLinkResourceListResult.
-func (p PrivateLinkResourceListResult) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", p.NextLink)
-	populate(objectMap, "value", p.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type PrivateLinkResourceProperties.
@@ -552,23 +489,23 @@ func (r RegistrationInfo) MarshalJSON() ([]byte, error) {
 func (r *RegistrationInfo) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", r, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "expirationTime":
-			err = unpopulateTimeRFC3339(val, &r.ExpirationTime)
+			err = unpopulateTimeRFC3339(val, "ExpirationTime", &r.ExpirationTime)
 			delete(rawMsg, key)
 		case "registrationTokenOperation":
-			err = unpopulate(val, &r.RegistrationTokenOperation)
+			err = unpopulate(val, "RegistrationTokenOperation", &r.RegistrationTokenOperation)
 			delete(rawMsg, key)
 		case "token":
-			err = unpopulate(val, &r.Token)
+			err = unpopulate(val, "Token", &r.Token)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", r, err)
 		}
 	}
 	return nil
@@ -586,20 +523,20 @@ func (r RegistrationInfoPatch) MarshalJSON() ([]byte, error) {
 func (r *RegistrationInfoPatch) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", r, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "expirationTime":
-			err = unpopulateTimeRFC3339(val, &r.ExpirationTime)
+			err = unpopulateTimeRFC3339(val, "ExpirationTime", &r.ExpirationTime)
 			delete(rawMsg, key)
 		case "registrationTokenOperation":
-			err = unpopulate(val, &r.RegistrationTokenOperation)
+			err = unpopulate(val, "RegistrationTokenOperation", &r.RegistrationTokenOperation)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", r, err)
 		}
 	}
 	return nil
@@ -622,14 +559,6 @@ func (r ResourceModelWithAllowedPropertySet) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ResourceProviderOperationList.
-func (r ResourceProviderOperationList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", r.NextLink)
-	populate(objectMap, "value", r.Value)
-	return json.Marshal(objectMap)
-}
-
 // MarshalJSON implements the json.Marshaller interface for type ScalingPlan.
 func (s ScalingPlan) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -646,14 +575,6 @@ func (s ScalingPlan) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "systemData", s.SystemData)
 	populate(objectMap, "tags", s.Tags)
 	populate(objectMap, "type", s.Type)
-	return json.Marshal(objectMap)
-}
-
-// MarshalJSON implements the json.Marshaller interface for type ScalingPlanList.
-func (s ScalingPlanList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", s.NextLink)
-	populate(objectMap, "value", s.Value)
 	return json.Marshal(objectMap)
 }
 
@@ -715,13 +636,6 @@ func (s ScalingSchedule) MarshalJSON() ([]byte, error) {
 	return json.Marshal(objectMap)
 }
 
-// MarshalJSON implements the json.Marshaller interface for type ServiceSpecification.
-func (s ServiceSpecification) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "logSpecifications", s.LogSpecifications)
-	return json.Marshal(objectMap)
-}
-
 // MarshalJSON implements the json.Marshaller interface for type SessionHostHealthCheckFailureDetails.
 func (s SessionHostHealthCheckFailureDetails) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]interface{})
@@ -735,34 +649,26 @@ func (s SessionHostHealthCheckFailureDetails) MarshalJSON() ([]byte, error) {
 func (s *SessionHostHealthCheckFailureDetails) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "errorCode":
-			err = unpopulate(val, &s.ErrorCode)
+			err = unpopulate(val, "ErrorCode", &s.ErrorCode)
 			delete(rawMsg, key)
 		case "lastHealthCheckDateTime":
-			err = unpopulateTimeRFC3339(val, &s.LastHealthCheckDateTime)
+			err = unpopulateTimeRFC3339(val, "LastHealthCheckDateTime", &s.LastHealthCheckDateTime)
 			delete(rawMsg, key)
 		case "message":
-			err = unpopulate(val, &s.Message)
+			err = unpopulate(val, "Message", &s.Message)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type SessionHostList.
-func (s SessionHostList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", s.NextLink)
-	populate(objectMap, "value", s.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type SessionHostPatch.
@@ -802,76 +708,68 @@ func (s SessionHostProperties) MarshalJSON() ([]byte, error) {
 func (s *SessionHostProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "agentVersion":
-			err = unpopulate(val, &s.AgentVersion)
+			err = unpopulate(val, "AgentVersion", &s.AgentVersion)
 			delete(rawMsg, key)
 		case "allowNewSession":
-			err = unpopulate(val, &s.AllowNewSession)
+			err = unpopulate(val, "AllowNewSession", &s.AllowNewSession)
 			delete(rawMsg, key)
 		case "assignedUser":
-			err = unpopulate(val, &s.AssignedUser)
+			err = unpopulate(val, "AssignedUser", &s.AssignedUser)
 			delete(rawMsg, key)
 		case "friendlyName":
-			err = unpopulate(val, &s.FriendlyName)
+			err = unpopulate(val, "FriendlyName", &s.FriendlyName)
 			delete(rawMsg, key)
 		case "lastHeartBeat":
-			err = unpopulateTimeRFC3339(val, &s.LastHeartBeat)
+			err = unpopulateTimeRFC3339(val, "LastHeartBeat", &s.LastHeartBeat)
 			delete(rawMsg, key)
 		case "lastUpdateTime":
-			err = unpopulateTimeRFC3339(val, &s.LastUpdateTime)
+			err = unpopulateTimeRFC3339(val, "LastUpdateTime", &s.LastUpdateTime)
 			delete(rawMsg, key)
 		case "osVersion":
-			err = unpopulate(val, &s.OSVersion)
+			err = unpopulate(val, "OSVersion", &s.OSVersion)
 			delete(rawMsg, key)
 		case "objectId":
-			err = unpopulate(val, &s.ObjectID)
+			err = unpopulate(val, "ObjectID", &s.ObjectID)
 			delete(rawMsg, key)
 		case "resourceId":
-			err = unpopulate(val, &s.ResourceID)
+			err = unpopulate(val, "ResourceID", &s.ResourceID)
 			delete(rawMsg, key)
 		case "sessionHostHealthCheckResults":
-			err = unpopulate(val, &s.SessionHostHealthCheckResults)
+			err = unpopulate(val, "SessionHostHealthCheckResults", &s.SessionHostHealthCheckResults)
 			delete(rawMsg, key)
 		case "sessions":
-			err = unpopulate(val, &s.Sessions)
+			err = unpopulate(val, "Sessions", &s.Sessions)
 			delete(rawMsg, key)
 		case "status":
-			err = unpopulate(val, &s.Status)
+			err = unpopulate(val, "Status", &s.Status)
 			delete(rawMsg, key)
 		case "statusTimestamp":
-			err = unpopulateTimeRFC3339(val, &s.StatusTimestamp)
+			err = unpopulateTimeRFC3339(val, "StatusTimestamp", &s.StatusTimestamp)
 			delete(rawMsg, key)
 		case "sxSStackVersion":
-			err = unpopulate(val, &s.SxSStackVersion)
+			err = unpopulate(val, "SxSStackVersion", &s.SxSStackVersion)
 			delete(rawMsg, key)
 		case "updateErrorMessage":
-			err = unpopulate(val, &s.UpdateErrorMessage)
+			err = unpopulate(val, "UpdateErrorMessage", &s.UpdateErrorMessage)
 			delete(rawMsg, key)
 		case "updateState":
-			err = unpopulate(val, &s.UpdateState)
+			err = unpopulate(val, "UpdateState", &s.UpdateState)
 			delete(rawMsg, key)
 		case "virtualMachineId":
-			err = unpopulate(val, &s.VirtualMachineID)
+			err = unpopulate(val, "VirtualMachineID", &s.VirtualMachineID)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type StartMenuItemList.
-func (s StartMenuItemList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", s.NextLink)
-	populate(objectMap, "value", s.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type SystemData.
@@ -890,43 +788,35 @@ func (s SystemData) MarshalJSON() ([]byte, error) {
 func (s *SystemData) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", s, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "createdAt":
-			err = unpopulateTimeRFC3339(val, &s.CreatedAt)
+			err = unpopulateTimeRFC3339(val, "CreatedAt", &s.CreatedAt)
 			delete(rawMsg, key)
 		case "createdBy":
-			err = unpopulate(val, &s.CreatedBy)
+			err = unpopulate(val, "CreatedBy", &s.CreatedBy)
 			delete(rawMsg, key)
 		case "createdByType":
-			err = unpopulate(val, &s.CreatedByType)
+			err = unpopulate(val, "CreatedByType", &s.CreatedByType)
 			delete(rawMsg, key)
 		case "lastModifiedAt":
-			err = unpopulateTimeRFC3339(val, &s.LastModifiedAt)
+			err = unpopulateTimeRFC3339(val, "LastModifiedAt", &s.LastModifiedAt)
 			delete(rawMsg, key)
 		case "lastModifiedBy":
-			err = unpopulate(val, &s.LastModifiedBy)
+			err = unpopulate(val, "LastModifiedBy", &s.LastModifiedBy)
 			delete(rawMsg, key)
 		case "lastModifiedByType":
-			err = unpopulate(val, &s.LastModifiedByType)
+			err = unpopulate(val, "LastModifiedByType", &s.LastModifiedByType)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", s, err)
 		}
 	}
 	return nil
-}
-
-// MarshalJSON implements the json.Marshaller interface for type UserSessionList.
-func (u UserSessionList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", u.NextLink)
-	populate(objectMap, "value", u.Value)
-	return json.Marshal(objectMap)
 }
 
 // MarshalJSON implements the json.Marshaller interface for type UserSessionProperties.
@@ -945,32 +835,32 @@ func (u UserSessionProperties) MarshalJSON() ([]byte, error) {
 func (u *UserSessionProperties) UnmarshalJSON(data []byte) error {
 	var rawMsg map[string]json.RawMessage
 	if err := json.Unmarshal(data, &rawMsg); err != nil {
-		return err
+		return fmt.Errorf("unmarshalling type %T: %v", u, err)
 	}
 	for key, val := range rawMsg {
 		var err error
 		switch key {
 		case "activeDirectoryUserName":
-			err = unpopulate(val, &u.ActiveDirectoryUserName)
+			err = unpopulate(val, "ActiveDirectoryUserName", &u.ActiveDirectoryUserName)
 			delete(rawMsg, key)
 		case "applicationType":
-			err = unpopulate(val, &u.ApplicationType)
+			err = unpopulate(val, "ApplicationType", &u.ApplicationType)
 			delete(rawMsg, key)
 		case "createTime":
-			err = unpopulateTimeRFC3339(val, &u.CreateTime)
+			err = unpopulateTimeRFC3339(val, "CreateTime", &u.CreateTime)
 			delete(rawMsg, key)
 		case "objectId":
-			err = unpopulate(val, &u.ObjectID)
+			err = unpopulate(val, "ObjectID", &u.ObjectID)
 			delete(rawMsg, key)
 		case "sessionState":
-			err = unpopulate(val, &u.SessionState)
+			err = unpopulate(val, "SessionState", &u.SessionState)
 			delete(rawMsg, key)
 		case "userPrincipalName":
-			err = unpopulate(val, &u.UserPrincipalName)
+			err = unpopulate(val, "UserPrincipalName", &u.UserPrincipalName)
 			delete(rawMsg, key)
 		}
 		if err != nil {
-			return err
+			return fmt.Errorf("unmarshalling type %T: %v", u, err)
 		}
 	}
 	return nil
@@ -992,14 +882,6 @@ func (w Workspace) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "systemData", w.SystemData)
 	populate(objectMap, "tags", w.Tags)
 	populate(objectMap, "type", w.Type)
-	return json.Marshal(objectMap)
-}
-
-// MarshalJSON implements the json.Marshaller interface for type WorkspaceList.
-func (w WorkspaceList) MarshalJSON() ([]byte, error) {
-	objectMap := make(map[string]interface{})
-	populate(objectMap, "nextLink", w.NextLink)
-	populate(objectMap, "value", w.Value)
 	return json.Marshal(objectMap)
 }
 
@@ -1054,9 +936,12 @@ func populateByteArray(m map[string]interface{}, k string, b []byte, f runtime.B
 	}
 }
 
-func unpopulate(data json.RawMessage, v interface{}) error {
+func unpopulate(data json.RawMessage, fn string, v interface{}) error {
 	if data == nil {
 		return nil
 	}
-	return json.Unmarshal(data, v)
+	if err := json.Unmarshal(data, v); err != nil {
+		return fmt.Errorf("struct field %s: %v", fn, err)
+	}
+	return nil
 }